requirements.txt
# ── LLM / Agent Framework ─────────────────────────────────────────────────
groq>=0.8.0               # Free LLaMA 3 API (script + prompts + story)
crewai>=0.55.0            # Multi-agent orchestration framework
langchain>=0.2.0          # Tool wrappers used by CrewAI

# ── Voice (Edge-TTS — completely free, Microsoft voices) ──────────────────
edge-tts>=6.1.9           # High-quality free TTS (no API key)
gTTS>=2.5.0               # Optional Google Translate TTS fallback with Indian accent tld

# ── Image generation ──────────────────────────────────────────────────────
requests>=2.31.0          # Pollinations.ai (free, no key)
Pillow>=10.0.0            # Image processing & thumbnail creation

# ── Video & Audio ─────────────────────────────────────────────────────────
moviepy>=1.0.3            # Video assembly
pydub>=0.25.1             # Audio manipulation
ffmpeg-python>=0.2.0      # FFmpeg bindings

# ── Captions (Whisper — open source, runs locally) ────────────────────────
openai-whisper>=20231117  # Speech-to-text for auto captions
# OR use faster-whisper for speed:
# faster-whisper>=0.10.0

# ── Backend API ───────────────────────────────────────────────────────────
fastapi>=0.111.0          # REST API
uvicorn>=0.30.0           # ASGI server
python-multipart>=0.0.9   # File uploads in FastAPI
websockets>=12.0          # Real-time progress updates

# ── Queue & State ─────────────────────────────────────────────────────────
redis>=5.0.0              # Job queue (optional, for production)
celery>=5.3.0             # Async task queue (optional)

# ── Upload (YouTube) ──────────────────────────────────────────────────────
google-api-python-client>=2.130.0  # YouTube Data API v3 (free)
google-auth-oauthlib>=1.2.0

# ── Utilities ─────────────────────────────────────────────────────────────
python-dotenv>=1.0.0      # .env file support
pydantic>=2.7.0           # Data validation
aiofiles>=23.2.1          # Async file I/O
httpx>=0.27.0             # Async HTTP client
colorama>=0.4.6           # Colored terminal output
tqdm>=4.66.0              # Progress bars
